More hook progress, working-ish example
The example (a button hook) only works because we chose a lucky
location for the hook. Right now if you hook SVC code (most of the
program seems to run in supervisor mode) you'll always trash lr,
since the caller has to save lr in this case. (Fuck ARM7.)

So. time to reconsider. either this means adding push/pop code to
the hook, or switching away from svc entirely.